issue255: move get_latest_version from client to utils, and amend tests
